Persist multi-session cookies across app sessions
Allows the webview to preserve persisted cookies (not session cookies) across Maker Toolkit app sessions.

Documentation on this one was a little tricky to track down, but here's the most relevant bits:

- [`<webview>` tag `partition` attribute](https://electronjs.org/docs/api/webview-tag#partition)
- A little more info on `partition` can be found in the [docs for `BrowserWindow`'s `webPreferences` option](https://electronjs.org/docs/api/browser-window#new-browserwindowoptions)
- [This discussion on persistent partitions](https://github.com/electron/electron/issues/959#issuecomment-230153682) was very helpful.

This should be all we need to support the "Remember me" feature of Code.org.  Unfortunately, while working on this I learned that "Remember me" is broken for all browsers right now!