Whitelist sso.pcsd1.org
Requested independently in two Zendesk tickets:

- https://codeorg.zendesk.com/agent/tickets/135552
- https://codeorg.zendesk.com/agent/tickets/134994

It would be preferable not to handle school-specific URLs here but we don't currently have a good workaround.

I'm not entirely sure this will solve the problem for this district, as I don't have details on their OAuth flow beyond the origin they supplied.  Rather than asking them to track the whole series of redirects involved from their end, I think we should release this quick fix and ask them to test it.  In the meantime we should set up metrics on whitelist misses that will make this easier to diagnose/fix from our end in the future.